add selenium tool
